dcnn: split dcnn.cpp into dcnn.c and caffe.cpp (minimize c++ exposure)
