+++ Makefile
@@ -141,7 +141,7 @@ INCLUDES=-I.
 
 OBJS=board.o gtp.o move.o ownermap.o pattern3.o pattern.o patternsp.o patternprob.o playout.o probdist.o random.o stone.o timeinfo.o network.o fbook.o chat.o
 ifdef DCNN
-	OBJS+=dcnn.o
+	OBJS+=dcnn.o caffe.o
 endif
 SUBDIRS=random replay patternscan patternplay joseki montecarlo uct uct/policy playout tactics t-unit distributed
